You no longer prematurely fall off ladders.
This was being caused by the fallController overriding the ladder
Controller. Now ladderController will communicate with fallController
and you wont fall if there is a ladder. The manner of communication is
sketchy at best and there is a frame of delay. It doesn't seem to be
noticable. fallController and ladderController purhaps should be
combined along with playerController into one giant script since they
basically all do the same thing.