Started on a crafting system.
Needs to be finished. Algorithm to determine whether or not you can
fufil a crafting recipe is horrible O(n^3) This needs to be fixed,
presumably with some sort of comparator. Other than that, code will now
give back a list of fufilled recipe number. Need to present that to
user, get input and then make a function that takes a recipe number and
subtracts the required reactants from inven and adds products.